requirements.txt
absl-py==2.1.0
accelerate==1.0.1
antlr4-python3-runtime==4.9.3
av==12.3.0
beartype==0.19.0
braceexpand==0.1.7
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
einops==0.8.0
einops-exts==0.0.4
ema-pytorch==0.7.6
filelock==3.16.1
flamingo-pytorch==0.1.2
fonttools==4.55.0
fsspec==2024.10.0
ftfy==6.2.3
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==1.0.0
grpcio==1.68.0
huggingface-hub==0.26.2
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
lmdb==1.5.1
lpips==0.1.4
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
packaging==21.3
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
prettytable==3.11.0
protobuf==5.28.3
psutil==6.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.1.4
pyrootutils==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
scipy==1.10.1
sentencepiece==0.2.0
sentry-sdk==2.19.0
setproctitle==1.3.4
six==1.16.0
smmap==5.0.1
sympy==1.13.3
tenacity==9.0.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
timm==0.9.10
tokenizers==0.15.2
torch==2.4.1
torchdata==0.8.0
torchtyping==0.1.5
torchvision==0.19.1
tqdm==4.67.1
transformers==4.38.2
triton==3.0.0
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.3
wandb==0.18.7
wcwidth==0.2.13
webdataset==0.2.100
Werkzeug==3.0.6
zipp==3.20.2
